+++ OLD.py
@@ -1,14 +1,14 @@
 import discord
 from discord import client
 from discord.ext import commands, tasks
-from discord_slash import SlashCommand
+# from discord_slash import SlashCommand
 import random
 import os
 
 client = commands.Bot(command_prefix="$")
 v = "v0.01"
 client.remove_command("help")
-slash = SlashCommand(client, sync_commands=True)
+# slash = SlashCommand(client, sync_commands=True)
 
 
 @client.event